+++ packages/planum/src/select/select-popup/select-popup.tsx
@@ -8,7 +8,7 @@ import { DesktopPopup } from './select-popup-desktop'
 import { MobilePopup } from './select-popup-mobile'
 
 export function SelectPopup(props: React.PropsWithChildren<SelectPopupProps>) {
-  const { children, popupCss } = props
+  const { children, popupCss, popupSearchCss } = props
   const { select, state } = useSelectContext()
   const isMobile = useMediaQuery('(max-width: 768px)')
   const Popup = isMobile ? MobilePopup : DesktopPopup
@@ -17,7 +17,11 @@ export function SelectPopup(props: React.PropsWithChildren<SelectPopupProps>) {
     <FloatingNode id={select.nodeId}>
       <FloatingPortal id="planum-portal">
         <AnimatePresence>
-          {state.isOpen && <Popup popupCss={popupCss}>{children}</Popup>}
+          {state.isOpen && (
+            <Popup popupCss={popupCss} popupSearchCss={popupSearchCss}>
+              {children}
+            </Popup>
+          )}
         </AnimatePresence>
       </FloatingPortal>
     </FloatingNode>